proposed-ordinance/requirements.txt
docx2txt
pyyaml
titlecase

